Time: 644 ms (7.89%), Space: 7.6 MB (98.66%) - LeetHub

+++ 1033-broken-calculator/1033-broken-calculator.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int brokenCalc(int start, int tar) {
-        long long curr=tar;
+        int curr=tar;
         int oper=0;
         while(curr!=start&&curr>=0){
             if(curr<start){
